Use typescript for jest tests
* Adds ts-jest for easy management of Typescript test specs. https://github.com/kulshekhar/ts-jest
* Rewrite initial test in ts.
