+++ package.json
@@ -57,6 +57,7 @@
     "ifc2x4_helper.d.ts"
   ],
   "devDependencies": {
+    "@types/jest": "^27.0.3",
     "@types/three": "^0.130.1",
     "concurrently": "^6.2.0",
     "cpy": "^8.1.2",
@@ -67,6 +68,7 @@
     "npm-watch": "^0.10.0",
     "serve": "^12.0.0",
     "three": "^0.130.1",
+    "ts-jest": "^27.0.7",
     "typescript": "^4.3.5"
   },
   "browser": {
